requirements.txt
pyserial==3.5
Flask==3.0.3
playsound==1.2.2
pygame==2.6.1
